+++ infra/modules/messages-dynamodb/outputs.tf
@@ -1,0 +1,9 @@
+output "table_arn" {
+  description = "ARN of the messages DynamoDB table"
+  value       = aws_dynamodb_table.messages.arn
+}
+
+output "table_name" {
+  description = "Name of the messages DynamoDB table"
+  value       = aws_dynamodb_table.messages.name
+}